+++ packages/polkadart_scale_codec/test/direct_tests/array_test.dart
@@ -57,14 +57,15 @@ void main() {
 
   group('Array Codec Decode Test', () {
     test('When value 0x01020304 is decoded then it returns [1, 2, 3, 4]', () {
-      final input = HexInput('0x01020304');
+      final input = Input.fromHex('0x01020304');
       final result = ArrayCodec(U8Codec.codec, 4).decode(input);
       expect(result, [1, 2, 3, 4]);
+      expect(input.remainingLength, 0);
     });
 
     test('When value 0x01020304 is decoded then it returns [[1, 2], [3, 4]]',
         () {
-      final input = HexInput('0x01020304');
+      final input = Input.fromHex('0x01020304');
       final result = ArrayCodec(
         ArrayCodec(U8Codec.codec, 2),
         2,
@@ -73,25 +74,28 @@ void main() {
         [1, 2],
         [3, 4]
       ]);
+      expect(input.remainingLength, 0);
     });
     test('When value 0x05060708 is decoded then it returns [5, 6, 7, 8]', () {
-      final input = HexInput('0x05060708');
+      final input = Input.fromHex('0x05060708');
       final result = ArrayCodec(U8Codec.codec, 4).decode(input);
       expect(result, [5, 6, 7, 8]);
+      expect(input.remainingLength, 0);
     });
 
     test(
         'When value 0x05060708090a0b0c0d0e0f1011121314 is decoded then it returns [5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20]',
         () {
-      final input = HexInput('0x05060708090a0b0c0d0e0f1011121314');
+      final input = Input.fromHex('0x05060708090a0b0c0d0e0f1011121314');
       final result = ArrayCodec(U8Codec.codec, 16).decode(input);
       expect(
           result, [5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20]);
+      expect(input.remainingLength, 0);
     });
     test(
         'When value 0x0001010002010300 is decoded then it returns [[0, true], [1, false], [2, true], [3, false]]',
         () {
-      final input = HexInput('0x0001010002010300');
+      final input = Input.fromHex('0x0001010002010300');
       final result = ArrayCodec(
         TupleCodec([U8Codec.codec, BoolCodec.codec]),
         4,
@@ -102,6 +106,7 @@ void main() {
         [2, true],
         [3, false]
       ]);
+      expect(input.remainingLength, 0);
     });
   });
 }