+++ packages/polkadart_scale_codec/test/direct_tests/composite_test.dart
@@ -59,7 +59,7 @@ void main() {
   group('Composite Decode Test', () {
     test('When value 0x2a01 is decoded then it returns {"a": 42, "b": true}',
         () {
-      final input = HexInput('0x2a01');
+      final input = Input.fromHex('0x2a01');
       final decoded = CompositeCodec(
         {
           'a': U8Codec.codec,
@@ -70,12 +70,13 @@ void main() {
         'a': 42,
         'b': true,
       });
+      expect(input.remainingLength, 0);
     });
 
     test(
         'When Composite is decoded then it returns correct value {"index": 8, "note": "This is a note", "Juice": {"name": "Kiwi", "ounces": 1}, "Remarks": "Hey Food was good"}',
         () {
-      final input = HexInput(
+      final input = Input.fromHex(
           '0x0838546869732069732061206e6f7465104b69776901014448657920466f6f642077617320676f6f64');
       final decoded = CompositeCodec(
         {
@@ -101,6 +102,7 @@ void main() {
             },
             'Remarks': 'Hey Food was good',
           }.toString());
+      expect(input.remainingLength, 0);
     });
   });
 }