+++ packages/scale_codec_generator/example/scale_codec_generator_example.dart
@@ -1,6 +1,0 @@
-import 'package:scale_codec_generator/scale_codec_generator.dart';
-
-void main() {
-  final awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}