+++ packages/polkadart_scale_codec/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.1.0-pre.1] - 2023-03-21
+
+### Prerelease version
+- [#283](https://github.com/rankanizer/polkadart/pull/283) Removed `HexInput`, now must use `Input.fromHex()`
+- [#283](https://github.com/rankanizer/polkadart/pull/283) Use `Uint8Buffer` instead `BytesBuilder` when enconding.
+
 ## [1.0.0] - 2022-08-24
 
 ### First Version